fix: Clean up CIRISManager routes mess - use single routes.py
- Renamed routes_v2.py to routes.py (the actual working version)
- Renamed old routes.py to routes_old.py
- Updated all imports to use the single routes.py
- Fixed deployment script to use proper CIRISManager instance

This ensures production gets the full CIRISManager functionality, not just
the minimal agent discovery API.
