+++ ciris_engine/protocols/services/README.md
@@ -1,42 +1,50 @@
 # Service Protocols
 
-This directory contains the protocol definitions for all 19 CIRIS services. Each protocol defines the exact interface that its corresponding service implementation must follow.
+This directory contains the protocol definitions for all 21 CIRIS services. Each protocol defines the exact interface that its corresponding service implementation must follow.
 
 ## Service Categories
 
-### Graph Services (5)
+### Graph Services (6)
 Everything is memory - all state flows through these services:
-- **[graph_memory.py](./graph_memory.py)** - Core memory operations (memorize, recall, forget)
-- **[graph_audit.py](./graph_audit.py)** - Unified audit (graph + file + hash chain)
-- **[graph_config.py](./graph_config.py)** - Configuration as versioned nodes
-- **[graph_telemetry.py](./graph_telemetry.py)** - Metrics and telemetry data
-- **[graph_gratitude.py](./graph_gratitude.py)** - Gratitude tracking and reciprocity
+- **[memory.py](./graph/memory.py)** - Core memory operations (memorize, recall, forget)
+- **[config.py](./graph/config.py)** - Configuration as versioned nodes
+- **[telemetry.py](./graph/telemetry.py)** - Metrics and telemetry data
+- **[audit.py](./graph/audit.py)** - Unified audit (graph + file + hash chain)
+- **[incident_management.py](./graph/incident_management.py)** - Incident tracking and management
+- **[tsdb_consolidation.py](./graph/tsdb_consolidation.py)** - Time-series data consolidation
 
-### Core Services (5)
-Essential operations that cannot be stored in the graph:
-- **[llm.py](./llm.py)** - Primary language model interactions
-- **[mock_llm.py](./mock_llm.py)** - Deterministic LLM for testing
-- **[tool.py](./tool.py)** - External tool execution
-- **[secrets.py](./secrets.py)** - Secrets management and filtering
-- **[runtime_control.py](./runtime_control.py)** - Runtime coordination
+### Runtime Services (3)
+Core runtime capabilities:
+- **[llm.py](./runtime/llm.py)** - Primary language model interactions
+- **[runtime_control.py](./runtime/runtime_control.py)** - Runtime coordination
+- **[scheduler.py](./lifecycle/scheduler.py)** - Task scheduling and deferrals
 
-### Infrastructure Services (4)
+### Infrastructure Services (7)
 System foundations:
-- **[time.py](./time.py)** - Centralized time operations (no datetime.now()!)
-- **[shutdown.py](./shutdown.py)** - Graceful shutdown coordination
-- **[initialization.py](./initialization.py)** - Service initialization sequencing
-- **[visibility.py](./visibility.py)** - API introspection and transparency
+- **[time.py](./lifecycle/time.py)** - Centralized time operations (no datetime.now()!)
+- **[shutdown.py](./lifecycle/shutdown.py)** - Graceful shutdown coordination
+- **[initialization.py](./lifecycle/initialization.py)** - Service initialization sequencing
+- **[authentication.py](./infrastructure/authentication.py)** - Authentication and authorization
+- **[resource_monitor.py](./infrastructure/resource_monitor.py)** - Resource monitoring and limits
+- **[database_maintenance.py](./infrastructure/database_maintenance.py)** - Database health and cleanup
+- **[secrets.py](./runtime/secrets.py)** - Secrets management and filtering
 
-### Authority Services (1)
-Human oversight:
-- **[wise_authority.py](./wise_authority.py)** - All WA operations (auth, provisioning, deferrals)
+### Governance Services (4)
+System governance and control:
+- **[wise_authority.py](./governance/wise_authority.py)** - Human oversight and deferrals
+- **[filter.py](./governance/filter.py)** - Message filtering and prioritization
+- **[visibility.py](./governance/visibility.py)** - API introspection and transparency
+- **[self_observation.py](./adaptation/self_observation.py)** - Identity variance and pattern analysis
 
-### Special Services (4)
-Unique capabilities:
-- **[self_configuration.py](./self_configuration.py)** - Active adaptation intelligence (20% safeguard)
-- **[adaptive_filter.py](./adaptive_filter.py)** - Message filtering and prioritization
-- **[task_scheduler.py](./task_scheduler.py)** - Future task scheduling and DEFER support
-- **[transaction_orchestrator.py](./transaction_orchestrator.py)** - Multi-service atomic operations
+### Tool Services (1)
+External tool integration:
+- **[tool.py](./runtime/tool.py)** - External tool execution and coordination
+
+## Bus Protocols (Not Services)
+
+These define interfaces for adapter-provided functionality, not core services:
+- **[communication.py](./governance/communication.py)** - Interface for adapter communication
+- **[wa_auth.py](./governance/wa_auth.py)** - WA authorization interfaces
 
 ## Protocol Rules
