+++ ciris_engine/schemas/README.md
@@ -1,38 +1,109 @@
-# CIRIS Schema Architecture v1
+# CIRIS Schema Architecture
 
 This directory contains the organized, typed schema definitions that replace all `Dict[str, Any]` usage in CIRIS. Every data structure is now properly typed with Pydantic models.
 
-## Schema Categories
+## Schema Categories (24 directories)
 
-### 📋 [Actions](./actions/) - Action Context and Parameters
-Schemas for the 10 action types:
-- **contexts.py**: Action-specific context (SpeakContext, ToolContext, etc.)
-- **parameters.py**: Action-specific parameters (SpeakParameters, ToolParameters, etc.)
+### Core Schema Directories
 
-### 🧮 [DMA](./dma/) - Decision Making Schemas
-Schemas for DMA evaluations and decisions:
-- **decisions.py**: DMA results and action selection schemas
+#### 📋 [Actions](./actions/) - Action Parameters
+- **parameters.py**: Typed parameters for all 10 action types
 
-### 🎓 [Faculties](./faculties/) - Faculty Assessment Schemas
-Schemas for specialized reasoning assessments:
-- **assessments.py**: Faculty evaluation results
+#### 🧮 [DMA](./dma/) - Decision Making Algorithms
+- **core.py**: Base DMA schemas
+- **decisions.py**: DMA evaluation results
+- **faculty.py**: Faculty integration schemas
+- **prompts.py**: DMA prompt templates
+- **results.py**: Action selection results
 
-### 🛡️ [Guardrails](./guardrails/) - Safety Check Schemas
-Schemas for guardrail results:
-- **results.py**: Guardrail evaluation outcomes
+#### 🎯 [Handlers](./handlers/) - Handler Operations
+- **context.py**: Handler execution context
+- **core.py**: Base handler schemas
+- **memory_schemas.py**: Memory operation schemas
+- **schemas.py**: Handler-specific schemas
 
-### 🎯 [Handlers](./handlers/) - Handler Operation Schemas
-Schemas for handler operations:
-- **schemas.py**: Handler context and results
+#### 🏢 [Services](./services/) - Service Schemas
+- **graph/**: Graph service schemas (memory, telemetry, audit, etc.)
+- **authority/**: Wise Authority schemas
+- **core/**: Core service schemas (runtime, secrets)
+- **infrastructure/**: Infrastructure service schemas
+- **lifecycle/**: Lifecycle service schemas
+- **special/**: Special service schemas (self_observation)
+- **graph_core.py**: Graph node and edge schemas
+- **operations.py**: Service operation schemas
 
-### 🏢 [Services](./services/) - Service Interaction Schemas
-Schemas for service calls:
-- **metadata.py**: Service call metadata (replaces Dict[str, Any])
-- **requests.py**: Typed service requests and responses
+#### 🧠 [Processors](./processors/) - Cognitive Processing
+- **cognitive.py**: Cognitive state definitions
+- **state.py**: State transition schemas
+- **main.py**: Main processor schemas
+- **dma.py**: DMA processor integration
 
-### 🧠 [States](./states/) - Cognitive State Schemas
-Schemas for agent states:
-- **cognitive.py**: The 6 cognitive states and transitions
+### Infrastructure Schema Directories
+
+#### 🔌 [Adapters](./adapters/) - External Interfaces
+- **cli.py**: CLI adapter schemas
+- **discord.py**: Discord adapter schemas
+- **registration.py**: Adapter registration schemas
+- **tools.py**: Tool execution schemas
+
+#### 🔐 [API](./api/) - API Request/Response
+- **agent.py**: Agent endpoints
+- **auth.py**: Authentication schemas
+- **runtime.py**: Runtime control schemas
+- **telemetry.py**: Telemetry endpoints
+- **wa.py**: Wise Authority endpoints
+
+#### 📊 [Telemetry](./telemetry/) - Monitoring
+- **core.py**: Telemetry data schemas
+- **collector.py**: Metric collection schemas
+
+#### 🔒 [Secrets](./secrets/) - Security
+- **core.py**: Secret management schemas
+- **service.py**: Secrets service schemas
+
+### Supporting Schema Directories
+
+#### 🎭 [Conscience](./conscience/) - Ethical Reasoning
+- **core.py**: Conscience evaluation schemas
+- **results.py**: Ethical assessment results
+
+#### ⚙️ [Config](./config/) - Configuration
+- **agent.py**: Agent configuration schemas
+- **essential.py**: Essential config schemas
+
+#### 📝 [Audit](./audit/) - Audit Trail
+- **core.py**: Audit entry schemas
+- **verification.py**: Audit verification schemas
+
+#### 🏗️ [Infrastructure](./infrastructure/) - System Infrastructure
+- **identity_variance.py**: Identity monitoring schemas
+- **behavioral_patterns.py**: Pattern analysis schemas
+- **oauth.py**: OAuth authentication schemas
+
+#### 💾 [Persistence](./persistence/) - Database
+- **core.py**: Database operation schemas
+- **tables.py**: Table definition schemas
+
+#### 🚌 [Buses](./buses/) - Message Bus Schemas
+
+#### 🌐 [Context](./context/) - Context Management
+
+#### 📦 [Data](./data/) - Data Structures
+
+#### 📋 [Formatters](./formatters/) - Output Formatting
+
+#### 📚 [Registries](./registries/) - Service Registries
+
+#### 🏃 [Runtime](./runtime/) - Runtime System
+- **core.py**: Core runtime schemas
+- **enums.py**: System enumerations
+- **memory.py**: Memory operation schemas
+- **system_context.py**: System context schemas
+
+#### 🛠️ [Utils](./utils/) - Utilities
+- **config_validator.py**: Configuration validation
+
+#### 🔧 [tools.py](./tools.py) - Tool Schemas
 
 ## Key Schema Replacements
 
@@ -49,7 +120,7 @@ metadata: Dict[str, Any] = {
 ### After (Typed Schema)
 ```python
 # GOOD - typed
-from ciris_engine.schemas.v1.services import ServiceMetadata
+from ciris_engine.schemas.services import ServiceMetadata
 
 metadata = ServiceMetadata(
     service_name="memory",
@@ -63,7 +134,7 @@ metadata = ServiceMetadata(
 1. **No Dict[str, Any]**: Every dict is now a typed Pydantic model
 2. **Validation**: All inputs are validated at the boundary
 3. **Serialization**: All schemas support JSON serialization
-4. **Evolution**: Schemas are versioned (v1) for future migration
+4. **Evolution**: Schemas support backward-compatible evolution
 5. **Documentation**: Every field has a description
 
 ## Common Patterns
@@ -95,7 +166,7 @@ class ActionContext(BaseModel):
 To migrate from Dict[str, Any]:
 
 1. Identify the dict usage
-2. Find or create appropriate schema in v1/
+2. Find or create appropriate schema in the relevant directory
 3. Replace dict creation with schema instantiation
 4. Update type hints throughout call chain
 5. Run mypy to verify
